+++ Server/farm-manager-api/src/main/java/com/smartfarm/api/dto/AiAnalysisDto.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class AiAnalysisDto {
-    private Integer analysisId;
+    private Long analysisId; // bigint -> giữ Long
     private Integer pBatchId;
     private String batchName;
     private String imagePath;
